statistics macros: Avoid unsigned integer comparisons against 0
CID #199900, #199913, #199923, #199934, #199941, #199964, #199998,
 #200032, #200049, #200074, #200096.

(cherry picked from commit 39fc05ec0ff2122dd79ec5582ac020c999c11c30)
(cherry picked from commit b497af00652cac0421f9e316bec4d5d12bb9ede3)

+++ statistics.h
@@ -208,7 +208,7 @@ extern gen_lock_t *stat_lock;
 		#define update_stat( _var, _n) \
 			do { \
 				if ( !((_var)->flags&STAT_IS_FUNC) ) {\
-					if (_n>=0) \
+					if ((long)(_n) >= 0L) \
 						atomic_add( _n, (_var)->u.val);\
 					else \
 						atomic_sub( -(_n), (_var)->u.val);\
